+++ src/engine/terminal/lua/api_functions/mod.rs
@@ -1,0 +1,5 @@
+pub mod lua_draw_api;
+pub mod lua_input_api;
+pub mod lua_logs_api;
+pub mod lua_memory_api;
+pub mod lua_palette_api;
