+++ src/engine/terminal/lua/auxiliary_functions/mod.rs
@@ -1,1 +1,2 @@
 pub mod draw_render;
+pub mod mouse_pos;
