--- src/engine/terminal/lua/functions/mod.rs
@@ -1,1 +1,0 @@
-pub mod lua_draw_api;
