+++ src/engine/terminal/lua/mod.rs
@@ -1,5 +1,4 @@
 pub mod setup_lua_functions;
 pub mod handle_lua_files;
-pub mod lua_auxiliary_functions;
-pub mod functions;
+pub mod api_functions;
 pub mod auxiliary_functions;
